Moves needsClose needsClosedBySystem and needsOpen to needs-actions (better fit)
Move getConnectionRelatedData into connection-actions and remove the export of the function (it is only accessed there)
